Add check labels to event attributes
Currently, only entity labels are added to the Alerta event's attributes.

With this change, both entity and check labels are added to the attributes.

If entity and check have the same label, the check label will be appended to the entity label with a forward slash separator.

Labels with names that contain periods or dollar signs are omitted because these characters are not allowed in Alerta attribute names.