replace periods and dollar signs in attribute keys with underscores
The handler adds the entity and check labels to Alerta event attributes. Alerta does not allow periods or dollar signs in attribute names and the API will throw an error if they are present.

To prevent these errors and allow the event to be triggered, periods and dollar signs in the entity and check label names are replaced with underscores.