Add check labels (#16)
* Add check labels to event attributes

Currently, only entity labels are added to the Alerta event's attributes.

With this change, both entity and check labels are added to the attributes.

If entity and check have the same label, the check label will be appended to the entity label with a forward slash separator.

Labels with names that contain periods or dollar signs are omitted because these characters are not allowed in Alerta attribute names.

* Add labels to example event

Added timestamp because the handler errors out without one.

Added labels to entity and check to test handler code changes.

* replace periods and dollar signs in attribute keys with underscores

The handler adds the entity and check labels to Alerta event attributes. Alerta does not allow periods or dollar signs in attribute names and the API will throw an error if they are present.

To prevent these errors and allow the event to be triggered, periods and dollar signs in the entity and check label names are replaced with underscores.